Fix selectors/app module-boundary violations
metabase/selectors/app.ts (shared) imported dashboard and query_builder
selectors. Move the feature-orchestrating app-shell selectors (app-bar / nav-bar
/ breadcrumb visibility, collection id) plus the router and embed-chrome
primitives into a new app-tier metabase/app/selectors.ts, leaving only generic
feature-free state reads in the shared module.

getIsNavbarOpen stays shared (read across tiers) and is decoupled from the
now-app-tier getIsAppBarVisible: its embedded force-open guard uses the proven
feature-free equivalent embedded && side_nav && (!top_nav || fullscreen). The
equivalence is verified against the real getIsAppBarVisible across all input
combinations in app/get-is-navbar-open.unit.spec.ts.
