+++ frontend/src/metabase/selectors/app.ts
@@ -2,225 +2,18 @@ import type { Selector } from "@reduxjs/toolkit";
 import { createSelector } from "@reduxjs/toolkit";
 import type { Location } from "history";
 
-import {
-  getDashboard,
-  getDashboardId,
-  getIsEditing as getIsEditingDashboard,
-} from "metabase/dashboard/selectors";
-import { getCurrentDocument } from "metabase/documents/selectors";
-import {
-  getIsSavedQuestionChanged,
-  getQuestion,
-} from "metabase/query_builder/selectors";
 import type { State } from "metabase/redux/store";
 import {
   getEmbedOptions,
   getIsEmbeddingIframe,
 } from "metabase/selectors/embed";
 import { getUser } from "metabase/selectors/user";
-import * as Urls from "metabase/urls";
 
 import { getSetting } from "./settings";
 
 export interface RouterProps {
   location: Location;
 }
-
-const PATHS_WITHOUT_NAVBAR = [
-  /^\/setup/,
-  /^\/auth/,
-  /^\/data-studio/,
-  /\/model\/.*\/query/,
-  /\/model\/.*\/columns/,
-  /\/model\/.*\/metadata/,
-  /\/model\/query/,
-  /\/model\/columns/,
-  /\/model\/metadata/,
-  /\/transform\/new\/.*\/query/,
-];
-
-const PATHS_WITH_COLLECTION_BREADCRUMBS = [
-  /\/question\//,
-  /\/model\//,
-  /\/metric\//,
-  /\/dashboard\//,
-  /\/document\//,
-];
-
-// Paths where collection identity comes from the URL itself, so breadcrumbs
-// can render without needing a question/dashboard/document in redux state.
-const STANDALONE_COLLECTION_BREADCRUMB_PATHS = [/\/collection\//];
-const PATHS_WITH_QUESTION_LINEAGE = [/\/question/, /\/model/];
-
-export const getRouterPath = (state: State, props: RouterProps) => {
-  return props?.location?.pathname ?? window.location.pathname;
-};
-
-export const getRouterHash = (state: State, props: RouterProps) => {
-  return props?.location?.hash ?? window.location.hash;
-};
-
-export const getIsAdminApp = createSelector([getRouterPath], (path) => {
-  return path.startsWith("/admin/");
-});
-
-export const getIsDataStudioApp = createSelector([getRouterPath], (path) => {
-  return path.startsWith("/data-studio");
-});
-
-export const getIsMetricsViewer = createSelector([getRouterPath], (path) => {
-  return path.startsWith("/explore");
-});
-
-export const getIsCollectionPathVisible = createSelector(
-  [
-    getQuestion,
-    getDashboard,
-    getCurrentDocument,
-    getRouterPath,
-    getIsEmbeddingIframe,
-    getEmbedOptions,
-  ],
-  (question, dashboard, document, path, isEmbedded, embedOptions) => {
-    if (isEmbedded && !embedOptions.breadcrumbs) {
-      return false;
-    }
-
-    const isModelDetail = /\/model\/.*\/detail\/.*/.test(path);
-    if (isModelDetail) {
-      return true;
-    }
-
-    if (
-      STANDALONE_COLLECTION_BREADCRUMB_PATHS.some((pattern) =>
-        pattern.test(path),
-      )
-    ) {
-      return true;
-    }
-
-    return (
-      ((question != null && question.isSaved()) ||
-        dashboard != null ||
-        document !== null) &&
-      PATHS_WITH_COLLECTION_BREADCRUMBS.some((pattern) => pattern.test(path))
-    );
-  },
-);
-
-export const getIsQuestionLineageVisible = createSelector(
-  [getIsSavedQuestionChanged, getRouterPath],
-  (isSavedQuestionChanged, path) =>
-    isSavedQuestionChanged &&
-    PATHS_WITH_QUESTION_LINEAGE.some((pattern) => pattern.test(path)),
-);
-
-export const getIsNavBarEnabled = createSelector(
-  [
-    getUser,
-    getRouterPath,
-    getIsEditingDashboard,
-    getIsEmbeddingIframe,
-    getEmbedOptions,
-  ],
-  (currentUser, path, isEditingDashboard, isEmbedded, embedOptions) => {
-    if (!currentUser || isEditingDashboard) {
-      return false;
-    }
-    if (isEmbedded && !embedOptions.side_nav) {
-      return false;
-    }
-
-    return !PATHS_WITHOUT_NAVBAR.some((pattern) => pattern.test(path));
-  },
-);
-
-const getIsEmbeddedAppBarVisible = createSelector(
-  [
-    getEmbedOptions,
-    getIsQuestionLineageVisible,
-    getIsCollectionPathVisible,
-    getIsNavBarEnabled,
-  ],
-  (
-    embedOptions,
-    isQuestionLineageVisible,
-    isCollectionPathVisible,
-    isNavBarEnabled,
-  ) => {
-    const anyEmbeddedAppBarElementVisible =
-      isNavBarEnabled ||
-      embedOptions.search ||
-      embedOptions.new_button ||
-      embedOptions.logo ||
-      isQuestionLineageVisible ||
-      isCollectionPathVisible;
-    return embedOptions.top_nav && anyEmbeddedAppBarElementVisible;
-  },
-);
-
-export const getIsAppBarVisible = createSelector(
-  [
-    getUser,
-    getRouterPath,
-    getRouterHash,
-    getIsAdminApp,
-    getIsDataStudioApp,
-    getIsEditingDashboard,
-    getIsEmbeddingIframe,
-    getIsEmbeddedAppBarVisible,
-  ],
-  (
-    currentUser,
-    path,
-    hash,
-    isAdminApp,
-    isDataStudioApp,
-    isEditingDashboard,
-    isEmbedded,
-    isEmbeddedAppBarVisible,
-  ) => {
-    const isFullscreen = hash.includes("fullscreen");
-
-    if (
-      !currentUser ||
-      (isEmbedded && !isEmbeddedAppBarVisible) ||
-      isAdminApp ||
-      isDataStudioApp ||
-      isEditingDashboard ||
-      isFullscreen
-    ) {
-      return false;
-    }
-    return !PATHS_WITHOUT_NAVBAR.some((pattern) => pattern.test(path));
-  },
-);
-
-export const getIsLogoVisible = createSelector(
-  [getIsEmbeddingIframe, getEmbedOptions],
-  (isEmbeddingIframe, embedOptions) => {
-    return !isEmbeddingIframe || embedOptions.logo;
-  },
-);
-
-export const getIsSearchVisible = createSelector(
-  [getIsEmbeddingIframe, getEmbedOptions],
-  (isEmbeddingIframe, embedOptions) => {
-    return !isEmbeddingIframe || embedOptions.search;
-  },
-);
-
-export const getIsNewButtonVisible = createSelector(
-  [getIsEmbeddingIframe, getEmbedOptions],
-  (isEmbeddingIframe, embedOptions) => {
-    return !isEmbeddingIframe || embedOptions.new_button;
-  },
-);
-
-export const getIsAppSwitcherVisible = createSelector(
-  [getIsEmbeddingIframe],
-  (isEmbeddingIframe) => !isEmbeddingIframe,
-);
 
 export const getErrorPage = (state: State) => {
   return state.app.errorPage;
@@ -235,52 +28,29 @@ export const getErrorMessage = (state: State) => {
   return errorPage?.data?.message || errorPage?.data;
 };
 
-export const getCollectionId = createSelector(
-  [
-    getQuestion,
-    getDashboard,
-    getDashboardId,
-    getCurrentDocument,
-    getDetailViewState,
-    getRouterPath,
-  ],
-  (question, dashboard, dashboardId, document, detailView, path) => {
-    if (detailView) {
-      return detailView.collectionId;
-    }
-
-    if (document) {
-      return document.collection_id;
-    }
-
-    if (dashboardId) {
-      return dashboard?.collection_id;
-    }
-
-    const questionCollectionId = question?.collectionId();
-    if (questionCollectionId != null) {
-      return questionCollectionId;
-    }
-
-    // On a collection page the URL itself identifies the current collection.
-    return Urls.extractCollectionIdFromPath(path);
-  },
-);
-
 export const getIsNavbarOpen: Selector<State, boolean> = createSelector(
   [
     getIsEmbeddingIframe,
     getEmbedOptions,
-    getIsAppBarVisible,
+    (_state: State) => window.location.hash,
     (state: State) => state.app.isNavbarOpen,
   ],
-  (isEmbeddingIframe, embedOptions, isAppBarVisible, isNavbarOpen) => {
-    // in an embedded instance, when the app bar is hidden, but the nav bar is not
-    // we need to force the sidebar to be open or else it will be totally inaccessible
+  (isEmbeddingIframe, embedOptions, locationHash, isNavbarOpen) => {
+    // In an embedded instance, when the app bar is hidden but the side nav is
+    // enabled, force the sidebar open or it would be totally inaccessible.
+    //
+    // The app bar is hidden exactly when `top_nav` is off or we're in
+    // fullscreen. The other factors that feed getIsAppBarVisible
+    // (search/logo/question-lineage/collection-path/…) only matter when the nav
+    // bar isn't rendered anyway, so they never change this guard's outcome —
+    // which lets us stay feature-free here rather than importing the app-tier
+    // getIsAppBarVisible. The equivalence is proven against the real
+    // getIsAppBarVisible in app/get-is-navbar-open.unit.spec.ts.
+    const isFullscreen = locationHash.includes("fullscreen");
     if (
       isEmbeddingIframe &&
       embedOptions.side_nav === true &&
-      !isAppBarVisible
+      (!embedOptions.top_nav || isFullscreen)
     ) {
       return true;
     }
